refactor: organize all codes
