+++ src/components/ClassroomAdminPanel.tsx
@@ -1,4 +1,5 @@
 import { useCallback, useEffect, useState } from 'react';
+import { useAuthedFetch } from '@/hooks/useAuthedFetch';
 import type { ComponentProps } from 'react';
 
 type Classroom = {
@@ -20,16 +21,7 @@ export default function ClassroomAdminPanel({ getAccessTokenSilently }: Props) {
   const [isSubmitting, setIsSubmitting] = useState(false);
   const [error, setError] = useState<string | null>(null);
 
-  const authedFetch = useCallback(async (path: string, init?: RequestInit) => {
-    const token = await getAccessTokenSilently();
-    return fetch(path, {
-      ...init,
-      headers: {
-        Authorization: `Bearer ${token}`,
-        ...(init?.headers ?? {}),
-      },
-    });
-  }, [getAccessTokenSilently]);
+  const authedFetch = useAuthedFetch(getAccessTokenSilently);
 
   const loadClassrooms = useCallback(async () => {
     setIsLoading(true);
