+++ src/components/StudentManagementPanel.tsx
@@ -3,6 +3,7 @@ import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
 import type { Resolver, SubmitHandler } from 'react-hook-form';
+import { useAuthedFetch } from '@/hooks/useAuthedFetch';
 import type { CurrentUser } from '../types/currentUser';
 
 type Classroom = {
@@ -84,19 +85,7 @@ export default function StudentManagementPanel({ currentUser, getAccessTokenSile
     },
   });
 
-  const authedFetch = useCallback(
-    async (path: string, init?: RequestInit) => {
-      const token = await getAccessTokenSilently();
-      return fetch(path, {
-        ...init,
-        headers: {
-          Authorization: `Bearer ${token}`,
-          ...(init?.headers ?? {}),
-        },
-      });
-    },
-    [getAccessTokenSilently],
-  );
+  const authedFetch = useAuthedFetch(getAccessTokenSilently);
 
   const activeClassroomId = useMemo(() => {
     return isAdmin ? selectedClassroomId : (currentUser.classroomId ?? '');
