+++ src/hooks/useClassroomPresetsData.ts
@@ -1,0 +1,33 @@
+import { useCallback } from 'react';
+import type { AuthedFetch } from '@/hooks/useAuthedFetch';
+import type {
+  LessonTypeListItem,
+  SubjectListItem,
+  TimeSlotListItem,
+} from '@/types/api';
+
+/**
+ * 1 教室の subject / lesson-type / time-slot をまとめて取得（週次プリセット取得と同系統）。
+ * `signal` によるキャンセルに対応。
+ */
+export function useClassroomPresetsData(authedFetch: AuthedFetch) {
+  return useCallback(
+    async (classroomId: string, signal?: AbortSignal) => {
+      const [sRes, lRes, tRes] = await Promise.all([
+        authedFetch(`/api/classrooms/${encodeURIComponent(classroomId)}/subjects`, { signal }),
+        authedFetch(`/api/classrooms/${encodeURIComponent(classroomId)}/lesson-types`, { signal }),
+        authedFetch(`/api/classrooms/${encodeURIComponent(classroomId)}/time-slots`, { signal }),
+      ]);
+      if (!sRes.ok || !lRes.ok || !tRes.ok) {
+        return { ok: false as const, subjects: [] as SubjectListItem[], lessonTypes: [] as LessonTypeListItem[], timeSlots: [] as TimeSlotListItem[] };
+      }
+      const [subjects, lessonTypes, timeSlots] = await Promise.all([
+        sRes.json() as Promise<SubjectListItem[]>,
+        lRes.json() as Promise<LessonTypeListItem[]>,
+        tRes.json() as Promise<TimeSlotListItem[]>,
+      ]);
+      return { ok: true as const, subjects, lessonTypes, timeSlots };
+    },
+    [authedFetch],
+  );
+}
