+++ src/pages/CalendarPage.tsx
@@ -12,6 +12,7 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@/components/ui/select';
+import { useAuthedFetch } from '@/hooks/useAuthedFetch';
 import { startOfWeekSunday } from '@/lib/calendarTime';
 import type { CurrentUser } from '@/types/currentUser';
 
@@ -82,19 +83,7 @@ export default function CalendarPage({
     return Array.from({ length: 7 }, (_, i) => weekStart.add(i, 'day'));
   }, [weekStart]);
 
-  const authedFetch = useCallback(
-    async (path: string, init?: RequestInit) => {
-      const token = await getAccessTokenSilently();
-      return fetch(path, {
-        ...init,
-        headers: {
-          Authorization: `Bearer ${token}`,
-          ...(init?.headers ?? {}),
-        },
-      });
-    },
-    [getAccessTokenSilently],
-  );
+  const authedFetch = useAuthedFetch(getAccessTokenSilently);
 
   useEffect(() => {
     if (!isAdmin) {
