+++ src/types/role.ts
@@ -1,0 +1,4 @@
+/**
+ * 業務上の3ロール（DB users.role と同値）。未設定・不明は `CurrentUser.role` で null 扱い。
+ */
+export type AppRole = 'admin' | 'manager' | 'staff';
